Move the class DualDiscretization into detail namespace and into the detail directory (#366)
